Add array, field access expressions, fix lvalues
Now all productions that should accept lvalues do. In the AST,
lvalues are now just Exp nodes. There is also a new Stm variant,
in the AST, ShortVarDecStm = [ids]:exp+ exp+, where the expressions
on the left hand side should all be ids (this will have to be
checked during weeding, it was a shift/shift conflict to specify
just ids during parsing).