Implement command-line interface in Main class
Implement a command-line interface which supports the flags
required for milestone 2 as well as most of the debugging
infrastructure we've developed. In the process I've done some
clean up and removed a bunch of commented out and unused code in
the Main class. I also got rid of the static variables because I
was finding it unclear why we had them while we were passing in
other things as parameters.

Everything here is open to discussion!