Create jarfile for portability
Build a golite.jar as part of the standard build process. Access it
using a shell script named run or a Windows batch script named
golite.bat.

The motivation for this was that the old run script couldn't easily be
modified to support command-line arguments (it relied on
automatically altering the relative path of a file being input) but
I think this is preferable in general.