TypeChecker: Support nested return statements
Return statements can be embedded within loops, blocks, switches,
etc. It would be possible to find the function return type by
iteratively getting node.parent() until we reach the function
declaration, but it's easier (and more efficient, for what little
that matters here) to just keep track of the return type with a
TypeChecker property.