Split compiler run code according to need
A lot of the configuration and specialized output code goes back
to golite.Main, and some code for staged execution is added to
test_programs.TestProgram, which is the only place that needs.
What's left in the Compiler class is the code to interface with
the lexer, parser, weeder, and type checker.