TypeChecker: fix check of && and || operators
Fixes programs/valid/teams/types/group10/alias_operators.go.
The problem was that any boolean type was accepted for arguments,
but the type of the output was always being set to bool. This is
correct for comparison operators (==, >, <, etc) but not boolean
connectives. Logical negation was already doing this the correct
way.